infrastructure separated by domain
