+++ packages/api/src/cache/cache.module.ts
@@ -1,6 +1,6 @@
 import { CacheModule as CachingModule } from '@nestjs/cache-manager';
 import { Module } from '@nestjs/common';
-import { CacheMetricsDecorator } from 'src/common/decorators/cache-metrics.decorator';
+import { CacheMetricsDecorator } from 'src/cache/infrastructure/cache-metrics.decorator';
 import { MetricsModule } from 'src/metrics/metrics.module';
 import { MetricsService } from 'src/metrics/metrics.service';
 import { CacheService } from './cache.service';
