+++ packages/api/src/city/city.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { CacheModule } from 'src/cache/cache.module';
 import { CacheAccessor } from 'src/cache/interfaces/cache-service.interface';
-import { CityProviderAdapter } from 'src/common/adapters/city-provider.adapter';
-import { CityProviderLoggingDecorator } from 'src/common/decorators/city-provider-logging.decorator';
-import { CityProviderCacheProxy } from 'src/common/proxies/city-validation-cache.proxy';
+import { CityProviderAdapter } from 'src/city/infrastructure/adapters/city-provider.adapter';
+import { CityProviderLoggingDecorator } from 'src/city/infrastructure/decorators/city-provider-logging.decorator';
+import { CityProviderCacheProxy } from 'src/city/infrastructure/proxies/city-provider-cache.proxy';
 import { LoggerModule } from 'src/logger/logger.module';
 import { LoggerService } from 'src/logger/logger.service';
 import { CityExistsConstraint } from './city-exists.constraint';
