+++ packages/api/src/city/providers/openweather.provider.ts
@@ -5,7 +5,7 @@ import { AxiosError } from 'axios';
 import { firstValueFrom } from 'rxjs';
 import { CityNotFoundException } from 'src/common/errors/city-not-found.error';
 import { ExternalApiException } from 'src/common/errors/external-api.error';
-import { CurrentOpenWeatherFetchDto } from 'src/weather/constants/types/current-weather-api.type';
+import { CurrentOpenWeatherFetchDto } from 'src/weather/types/current-weather-api.type';
 import { ChainableCityProvider } from '../interfaces/chainable-city.provider';
 
 @Injectable()
