+++ packages/api/src/subscriptions/dtos/create-subscription.dto.ts
@@ -1,5 +1,5 @@
 import { IsEmail, IsEnum, IsNotEmpty } from 'class-validator';
-import { CityExists } from 'src/common/validators/city-exists.validator';
+import { CityExists } from 'src/city/validators/city-exists.validator';
 import { NotificationsFrequencies } from 'src/notifications/constants/enums/notification-frequencies.enum';
 
 export class CreateSubscriptionDto {
