+++ packages/api/src/weather/infrastructure/cache/weather-cache-prefixes.enum.ts
